backend/requirements.txt
Django>=5.0,<5.1
psycopg2-binary
django-ninja
mysqlclient
django-cors-headers
django-allauth
django-rq
gunicorn
python-dotenv
